+++ source/keyboard_mouse.h
@@ -325,7 +325,7 @@ sc_type TextToSC(LPTSTR aText);
 vk_type TextToVK(LPTSTR aText, modLR_type *pModifiersLR = NULL, bool aExcludeThoseHandledByScanCode = false
 	, bool aAllowExplicitVK = true, HKL aKeybdLayout = GetKeyboardLayout(0));
 vk_type CharToVKAndModifiers(TCHAR aChar, modLR_type *pModifiersLR, HKL aKeybdLayout);
-vk_type TextToSpecial(LPTSTR aText, UINT aTextLength, KeyEventTypes &aEventTypem, modLR_type &aModifiersLR
+vk_type TextToSpecial(LPTSTR aText, size_t aTextLength, KeyEventTypes &aEventTypem, modLR_type &aModifiersLR
 	, bool aUpdatePersistent);
 
 #ifdef ENABLE_KEY_HISTORY_FILE
